split login form file
